// src/modules/notifications/shared/enums/notification-type.enum.ts
export enum NotificationType {
  ACCOUNT_OPENED = 'account_opened',
  FRAUD_DETECTED = 'fraud_detected',
  TRANSFER_FAILED = 'transfer_failed',
  ACCOUNT_CLOSED = 'account_closed',
  ACCOUNT_FROZEN = 'account_frozen',
  MONEY_DEPOSITED = 'money_deposited',
  MONEY_WITHDRAWN = 'money_withdraw',
  TRANSFER_CREDITED = 'transfer_credited',
  TRANSFER_DEBITED = 'transfer_debited',
}